backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.31.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Celery
celery[redis]==5.4.0
redis==5.1.0

# RSS Parsing
feedparser==6.0.11
httpx==0.27.0

# AI (Gemini)
google-generativeai==0.8.0

# De-duplication
simhash==2.1.2

# Utilities
python-dateutil==2.9.0
tenacity==9.0.0
